Hieu fix loi

+++ Vantix_HRM/src/main/java/poly/edu/vantix_hrm/repository/AttendanceRepository.java
@@ -34,5 +34,5 @@ public interface AttendanceRepository extends JpaRepository<Attendance, Integer>
     // Dùng cho Attendance Management
     // Query mới: Lấy Attendance REJECTED theo ID phòng ban và khoảng thời gian
     @Query("SELECT a FROM Attendance a " + "JOIN FETCH a.employee e " + "JOIN FETCH a.shift s " + "WHERE e.department.id = :departmentId " + "AND a.status = 'REJECTED' " + "AND a.workDate BETWEEN :startDate AND :endDate " + "ORDER BY e.id ASC, a.workDate DESC")
-    List<Attendance> findRejectedByDepartmentAndDateRange(@Param("departmentId") Integer departmentId, @Param("startDate") LocalDate startDate, @Param("endDate") LocalDate endDate);
+    List<Attendance> findRejectedByDepartmentAndDateRange(@Param("departmentId") Long departmentId, @Param("startDate") LocalDate startDate, @Param("endDate") LocalDate endDate);
 }
